The-Gods-Of-Programming-cuarto-semestre/Python/Clase01/requirements.txt
# ===============================================
# SISTEMA DE GESTIÓN DE USUARIOS - LAB UML 1.1
# ===============================================
# 
# Dependencias para aplicación CRUD con PostgreSQL
# Implementa patrón DAO con manejo robusto de excepciones
#

# === DEPENDENCIAS DE BASE DE DATOS ===
psycopg2-binary==2.9.7  # Driver PostgreSQL (versión binaria)

# === DEPENDENCIAS DE INTERFAZ ===  
colorama==0.4.6         # Colores en terminal (Windows compatible)

# === DEPENDENCIAS DE CONFIGURACIÓN ===
python-dotenv==1.0.0    # Carga de variables de entorno desde .env

# === DEPENDENCIAS DE DESARROLLO Y TESTING ===
pytest==7.4.0           # Framework de testing avanzado
pytest-cov==4.1.0       # Coverage para tests
unittest-xml-reporting==3.2.0  # Reportes XML para CI/CD

# === DEPENDENCIAS OPCIONALES ===
# black==23.7.0           # Formateador de código
# flake8==6.0.0           # Linter de código
# mypy==1.5.1             # Type checker estático

# === INFORMACIÓN ADICIONAL ===
# Python >= 3.8 requerido
# PostgreSQL >= 12 recomendado
# 
# Para instalar todas las dependencias:
# pip install -r requirements.txt
#
# Para instalar con dependencias de desarrollo:
# pip install -r requirements.txt pytest black flake8 mypy
